// packages/database/src/pg/models/stake_update.ts
/**
 * This file was automatically generated by knex
 * Do not modify this file manually
 */
import { Increment } from '../interfaces/common';
import * as commons from './common_update';

export interface PgStakeRewardUpdate extends commons.BaseRecordUpdate {
  startDate?: Date | null;
  endDate?: Date | null;
  tokenVestingDate?: Date | null;
  tokensToDistribute?: number | null | Increment;
  token?: string | null;
  totalStaked?: number | null | Increment;
  totalAirdropped?: number | null | Increment;
  status?: string | null;
}

export interface PgStakeUpdate extends commons.BaseRecordUpdate {
  member?: string | null;
  space?: string | null;
  token?: string | null;
  amount?: number | null | Increment;
  value?: number | null | Increment;
  weeks?: number | null | Increment;
  expiresAt?: Date | null;
  expirationProcessed?: boolean | null;
  orderId?: string | null;
  billPaymentId?: string | null;
  type?: string | null;
  customMetadata?: string | any | null;
}